refactor: move user repositories to its own folder
