+++ src/repositories/postgres/user/delete-user.js
@@ -1,4 +1,4 @@
-import { PostgresHelper } from '../../db/postgres/helper.js'
+import { PostgresHelper } from '../../../db/postgres/helper.js'
 
 export class PostgresDeleteUserRepository {
     async execute(userId) {
