+++ src/repositories/postgres/user/get-user-by-email.js
@@ -1,4 +1,4 @@
-import { PostgresHelper } from '../../db/postgres/helper.js'
+import { PostgresHelper } from '../../../db/postgres/helper.js'
 
 export class PostgresGetUserByEmailRepository {
     async execute(email) {
