+++ src/repositories/postgres/user/get-user-by-id.js
@@ -1,4 +1,4 @@
-import { PostgresHelper } from '../../db/postgres/helper.js'
+import { PostgresHelper } from '../../../db/postgres/helper.js'
 
 export class PostgresGetUserByIdRepository {
     async execute(userId) {
